I added the search engine.
The search engine was added via the search search form in the html file and linked to the js file where a new function was created to handle a search whenever a user enters a request in the search-form. The function was also called from outside all functions with a city input of New York as a default City to show on page load/refresh.